+++ processor/src/main/java/org/realityforge/arez/processor/Constants.java
@@ -27,6 +27,7 @@ final class Constants
 
   static final String POST_CONSTRUCT_ANNOTATION_CLASSNAME = "javax.annotation.PostConstruct";
   static final String INJECT_ANNOTATION_CLASSNAME = "javax.inject.Inject";
+  static final String SCOPE_ANNOTATION_CLASSNAME = "javax.inject.Scope";
   static final String DAGGER_MODULE_CLASSNAME = "dagger.Module";
 
   private Constants()