+++ processor/src/test/resources/expected/com/example/repository/CompleteRepositoryExampleRepositoryDaggerModule.java
@@ -1,0 +1,18 @@
+package com.example.repository;
+
+import dagger.Module;
+import dagger.Provides;
+import javax.annotation.Generated;
+import javax.annotation.Nonnull;
+import javax.inject.Singleton;
+
+@Generated("org.realityforge.arez.processor.ArezProcessor")
+@Module
+public interface CompleteRepositoryExampleRepositoryDaggerModule {
+  @Nonnull
+  @Singleton
+  @Provides
+  static CompleteRepositoryExampleRepository provideRepository(final Arez_CompleteRepositoryExampleRepository repository) {
+    return repository;
+  }
+}