+++ processor/src/test/resources/expected/com/example/repository/RepositoryWithMultipleCtorsRepositoryDaggerModule.java
@@ -8,7 +8,7 @@ import javax.inject.Singleton;
 
 @Generated("org.realityforge.arez.processor.ArezProcessor")
 @Module
-public interface RepositoryWithMultipleCtorsDaggerModule {
+public interface RepositoryWithMultipleCtorsRepositoryDaggerModule {
   @Nonnull
   @Singleton
   @Provides