+++ processor/src/test/resources/input/com/example/inject/ScopedInjectModel.java
@@ -1,0 +1,19 @@
+package com.example.inject;
+
+import javax.inject.Singleton;
+import org.realityforge.arez.annotations.Action;
+import org.realityforge.arez.annotations.ArezComponent;
+
+@Singleton
+@ArezComponent
+public class ScopedInjectModel
+{
+  public ScopedInjectModel()
+  {
+  }
+
+  @Action
+  public void myActionStuff()
+  {
+  }
+}